feat: add delete confrim task modal

+++ src/components/UI/Task/Task.tsx
@@ -1,9 +1,11 @@
-import React, { FC } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 import { ITask } from '../../../types/boardTypes';
 import cl from './Task.module.scss';
 import { useDeleteTaskMutation } from '../../../API/tasksCalls';
 import { Draggable } from 'react-beautiful-dnd';
 import { useTranslate } from '../../../hooks/useTranslate';
+import { Modal } from '../../Modal/modal';
+import ConfirmModal from '../../Modal/modals/confirmModal';
 
 interface ITaskProps {
   task: ITask;
@@ -14,12 +16,32 @@ interface ITaskProps {
 
 const Task: FC<ITaskProps> = ({ task, boardId, columnId, index }) => {
   const { title, _id: id } = task;
-  const [deleteTask, {}] = useDeleteTaskMutation();
+  const [deleteTask, { isLoading }] = useDeleteTaskMutation();
   const T = useTranslate();
-
+  //Modal manipulations
+  ////////////////////////
+  const [isModalOpen, setModalOpen] = useState(false);
+  useEffect(() => {
+    if (!isLoading) {
+      setModalOpen(false);
+    }
+  }, [isLoading]);
+  ///////////////////////////////////////////////////
   const deleteTaskOnClick = () => {
-    deleteTask({ boardId, columnId, id });
+    setModalOpen(true);
+    // deleteTask({ boardId, columnId, id });
   };
+
+  function confirmDeleteTask(e: React.MouseEvent<HTMLElement>) {
+    const target = (e.target as HTMLElement).closest('input');
+    const value = target?.value;
+    if (value == 'No' || value == 'Нет') {
+      setModalOpen(false);
+    }
+    if (value == 'Yes' || value == 'Да') {
+      deleteTask({ boardId, columnId, id });
+    }
+  }
 
   return (
     <Draggable draggableId={id} index={index}>
@@ -32,6 +54,11 @@ const Task: FC<ITaskProps> = ({ task, boardId, columnId, index }) => {
         >
           <h3>{title}</h3>
           <button onClick={deleteTaskOnClick}>{T('Task.deleteTask')}</button>
+          {isModalOpen && (
+            <Modal>
+              <ConfirmModal handler={confirmDeleteTask} isLoading={isLoading}></ConfirmModal>
+            </Modal>
+          )}
         </div>
       )}
     </Draggable>
